requirements.txt
socket
paho-mqtt
IPython
